Task 14. 50 Fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,17 +8,17 @@
 int main(void)
 {
 	int i;
-	long long int n1, n2, sum;
+	long int n1, n2, sum;
 
 	n1 = 1;
 	n2 = 2;
 	sum = n1 + n2;
-	printf("%lld, %lld, ", n1, n2);
+	printf("%ld, %ld, ", n1, n2);
 
 	i = 3;
 	while (i <= 50)
 	{
-		printf("%lld, ", sum);
+		printf("%ld, ", sum);
 		n1 = n2;
 		n2 = sum;
 		sum = n1 + n2;
